Moved to a single mainwindow
Removed transition slots
Renamed transition slots when possible to be generic as to the origin
Fixed type error in gumbo-parser
Moved UI transition connection making into the transitions themselves as the various UI widgets are no longer persistent. They are owned by the mainwindow when they are the central widget and deleted when replaced.