Fixes for iOS fix breaking Android
QSize in Android and iOS seems to mean very different things when applied to the QListWidget::gridSize. Since we're content with the Android layout I'm leaving it as is. A preprocessor directive switches using QListWidget::setGridSize based on if IOS_SPECIFIC is set. If it is not a Null QSize is set instead.